correction bug décalage dans la position actuel

+++ projet-2012-54/src/Audio/IPlayer.java
@@ -9,5 +9,5 @@ public interface IPlayer
 	public float getPosition();
 	public void setPosition(float position);
 	public float getCurrentVolume();
-	public int getLength();
+	public float getLength();
 }
